Added back func to ncurses_ui and now when you press yes it do stuff, if you press no it don't do  stuff

+++ src/command_line_ui/GameLine.cpp
@@ -67,8 +67,10 @@ int GameLine::enter()
                 break;
             case ENTER:
                 delwin(sure);
-                bluise_core::delete_game(game->get_name());
-                bluise_core::saveGLL();
+                if(yes) {
+                    bluise_core::delete_game(game->get_name());
+                    bluise_core::saveGLL();
+                }
                 return 1;
             }
         }
@@ -83,7 +85,7 @@ int GameLine::enter()
         mvwprintw(info, 3, 2, string("Executable: " + game->get_executable()).c_str());
         mvwprintw(info, 4, 2, string("Save path: " + game->get_save_path()).c_str());
         wattron(info, A_STANDOUT);
-        mvwprintw(info, getmaxy(info) - 2, getmaxx(info) / 2 - 6, "Back");
+        mvwprintw(info, getmaxy(info) - 2, (getmaxx(info) - 4) / 2, "Back");
         wattroff(info, A_STANDOUT);
         char ch;
         while(ch = wgetch(info)) {
@@ -95,23 +97,49 @@ int GameLine::enter()
             }
         }
     }
-    // else if(menu.button()==Menu::BACK) {
-    //     WINDOW* ok = newwin(getmaxy(stdscr) / 2 , getmaxx(stdscr) / 2, getmaxy(stdscr) * 1 / 4, getmaxx(stdscr) * 1 / 4);
-    //     refresh();
-    //     box(ok, 0, 0);
-    //     mvwprintw(ok, getmaxy(ok) / 2, getmaxx(ok) / 2 - game.get_name().length() - 2, game.get_name().c_str());
-    //     wattron(ok, A_STANDOUT);
-    //     mvwprintw(ok, getmaxy(ok) - 2, getmaxx(ok) / 2 - 6, "Back");
-    //     wattroff(ok, A_STANDOUT);
-    //     char ch;
-    //     while(ch = wgetch(ok)) {
-    //         switch(ch) 
-    //         {
-    //         case ENTER:
-    //             delwin(ok);
-    //             return 2;
-    //         }
-    //     }
-    // }
+    else if(menu.button()==Menu::BACK) {
+        clear();
+        WINDOW* sure = newwin(getmaxy(stdscr) / 2 , getmaxx(stdscr) / 2, getmaxy(stdscr) * 1 / 4, getmaxx(stdscr) * 1 / 4);
+        refresh();
+        box(sure, 0, 0);
+        string ok = "Do you want to back up saves for the "+game->get_name()+"?";
+        mvwprintw(sure, 1, (getmaxx(sure) - ok.length()) / 2, ok.c_str());
+        wattron(sure, A_STANDOUT);
+        mvwprintw(sure, getmaxy(sure) - 2, getmaxx(sure) * 1 / 4, "YES");
+        wattroff(sure, A_STANDOUT);
+        mvwprintw(sure, getmaxy(sure) - 2, getmaxx(sure) * 3 / 4 - 1, "NO");
+        bool yes = true;
+        char ch;
+        while(ch = wgetch(sure)) {
+            switch(ch) 
+            {
+            case RIGHT_ARROW:
+            case LEFT_ARROW:
+            case 'a':
+            case 'd':    
+                if(yes) {
+                    mvwprintw(sure, getmaxy(sure) - 2, getmaxx(sure) * 1 / 4, "YES");
+                    wattron(sure, A_STANDOUT);
+                    mvwprintw(sure, getmaxy(sure) - 2, getmaxx(sure) * 3 / 4 - 1, "NO");
+                    wattroff(sure, A_STANDOUT);
+                    yes = false;
+                }
+                else {
+                    wattron(sure, A_STANDOUT);
+                    mvwprintw(sure, getmaxy(sure) - 2, getmaxx(sure) * 1 / 4, "YES");
+                    wattroff(sure, A_STANDOUT);
+                    mvwprintw(sure, getmaxy(sure) - 2, getmaxx(sure) * 3 / 4 - 1, "NO");
+                    yes = true;
+                }
+                break;
+            case ENTER:
+                delwin(sure);
+                if(yes) {
+                    bluise_core::back(game->get_name());
+                }
+                return 0;
+            }
+        }
+    }
 }
